pyproject.toml
[project]
name = "weather-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuri Osadchyi",email = "aurum.yo@gmail.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
django-dynamic-fixture = "^4.0.1"
